Add files via upload
A simple and easy-to-use calendar app built purely using Kotlin. If you were looking to learn Kotlin, then this is probably one of the best ways to start.
The goal of this app is quite simple which makes it perfect to get your hands dirty learning a completely new language for developing Android apps. Another cool thing you can learn from this project is to make custom desktop widgets for Android.

A simple calendar with events and a customizable widget.

A simple calendar with optional CalDAV synchronization. You can easily create recurring events and setup reminders, it can also display week numbers.

Contains a monthly view and an event list widget where you can customize the color of the text, as well as the alpha and the color of the background.

Contains no ads or unnecessary permissions. It is fully opensource, provides customizable colors.

The Storage permission is needed only for exporting or importing events from .ics files.

The Contacts permission is used only at importing contact birthdays and anniversaries.